% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emSRM.R
\name{emSRM}
\alias{emSRM}
\title{Robust EM algorithm for Spline Regression Mixture Model (SRM)}
\usage{
emSRM(
  X,
  Y,
  splineOrder,
  splineType = c("spline", "B-spline"),
  nknots,
  verbose = TRUE
)
}
\arguments{
\item{X}{observation times, long form [n*m, 1], i.e. n subjects each with m observations, observation time can be irregular}

\item{Y}{observed outcomes, wide form [n, m]}

\item{splineOrder}{B-Spline order}

\item{splineType}{select between c("spline", "B-spline") indicating either natural cubic spline or b-splines}

\item{nknots}{number of internal knots selected equally}

\item{verbose}{Screen output?}
}
\description{
Robust EM algorithm for Spline Regression Mixture Model (SRM)
}
