% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emSRMir.R
\name{emSRMir}
\alias{emSRMir}
\title{Robust EM algorithm for Spline Regression Mixture Model (SRM) with Irregular observations}
\usage{
emSRMir(
  X,
  Y,
  id,
  splineOrder,
  splineType = c("spline", "B-spline"),
  nknots,
  verbose = TRUE,
  MaxIter = 1000
)
}
\arguments{
\item{X}{observation times, long form [n*m, 1], i.e. n subjects each with m observations, observation time can be irregular}

\item{Y}{observed outcomes, long form [n*m, 1]}

\item{id}{subject indicators to tell which data belongs to which subject [n*m, 1]}

\item{splineOrder}{B-Spline order}

\item{splineType}{select between c("spline", "B-spline") indicating either natural cubic spline or b-splines}

\item{nknots}{number of internal knots selected equally}

\item{verbose}{Screen output?}

\item{MaxIter}{Maximum iteration}
}
\description{
Robust EM algorithm for Spline Regression Mixture Model (SRM) with Irregular observations
}
\examples{
test = generateToyDataset(n1 = 40, n2 = 30, n3 = 30)
X  = rep(test$x, 100)
id = rep(seq(100), each = 100)
Y  = matrix(t(test$Y), ncol = 1)
output = emSRMir(X, Y, id, 3, splineType = "B-spline", 3)
}
