% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateToyDataset.R
\name{generateToyDataset}
\alias{generateToyDataset}
\title{Simulate an example data with 100 observations for each trajectory}
\usage{
generateToyDataset(n1 = 40, n2 = 30, n3 = 30)
}
\arguments{
\item{n1, n2, n3}{number of subjects for each group}
}
\value{
x is the observation time points; Y is the data saved in wide form; Z stores the labels
}
\description{
Simulate an example data with 100 observations for each trajectory
}
