% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBreimanWaves.R
\name{sampleBreimanWaves}
\alias{sampleBreimanWaves}
\title{Simulate n Breiman Waves}
\usage{
sampleBreimanWaves(n)
}
\arguments{
\item{n}{number of waves to be simulated}
}
\value{
Y is the data; Z is the label
}
\description{
Simulate n Breiman Waves
}
